refactor: removed unnecessary items; packaged items.
