+++ app/src/main/java/team2/kakigowhere/ui/map/InfoWindowAdapter.kt
@@ -1,4 +1,4 @@
-package team2.kakigowhere.ui
+package team2.kakigowhere.ui.map
 
 import android.content.Context
 import android.graphics.Bitmap
