+++ app/src/main/java/team2/kakigowhere/ui/place/DetailFragment.kt
@@ -1,4 +1,4 @@
-package team2.kakigowhere.ui
+package team2.kakigowhere.ui.place
 
 import android.content.Context
