+++ app/src/main/java/team2/kakigowhere/ui/place/ItinerarySheetAdapter.kt
@@ -1,4 +1,4 @@
-package team2.kakigowhere.ui
+package team2.kakigowhere.ui.place
 
 import android.util.Log
 import android.view.LayoutInflater
