--- app/src/main/res/layout/activity_test.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:gravity="top|center"
-    android:padding="20dp">
-
-    <Button
-        android:id="@+id/test_api"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Test API call"/>
-
-</LinearLayout>
