+++ app/src/main/res/layout/fragment_itinerary.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.ItineraryFragment">
+    tools:context=".ui.itinerary.ItineraryFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
