+++ app/src/main/res/layout/fragment_itinerary_detail.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.ItineraryDetailFragment">
+    tools:context=".ui.itinerary.ItineraryDetailFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
