+++ app/src/main/res/layout/fragment_itinerary_edit.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".ui.ItineraryEditFragment">
+    tools:context=".ui.itinerary.ItineraryEditFragment">
 
     <TextView
         android:id="@+id/id_title"
@@ -26,7 +26,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="30dp"
-        android:textSize="26dp"
+        android:textSize="26sp"
         android:textStyle="bold"
         android:text="Notes"/>
