+++ app/src/main/res/layout/fragment_maps.xml
@@ -4,7 +4,7 @@
     android:id="@+id/fragment_maps_root"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.MapsFragment">
+    tools:context=".ui.map.MapsFragment">
 
     <!-- Google Map -->
     <fragment
