+++ app/src/main/res/layout/item_place.xml
