--- app/src/main/res/layout/item_place_explore.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="@color/material_dynamic_neutral95"
-    android:orientation="horizontal"
-    android:padding="8dp"
-    android:layout_marginBottom="8dp">
-
-    <ImageView
-        android:id="@+id/imgPlace"
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:src="@drawable/marina_bay_sands"
-        android:scaleType="centerCrop" />
-
-    <LinearLayout
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:orientation="vertical"
-        android:paddingStart="8dp">
-
-        <TextView
-            android:id="@+id/tvName"
-            android:text="Place Name"
-            android:textStyle="bold"
-            android:textSize="16sp"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:maxLines="2"
-            android:ellipsize="end"
-            />
-
-        <TextView
-            android:id="@+id/tvRating"
-            android:text="4.5 ★★★★☆"
-            android:textSize="14sp"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"/>
-
-    </LinearLayout>
-</LinearLayout>
